:sparkles: feat: RootInstanceをシングルトン化とFixedLoadSceneAreaをLoader利用に修正
