+++ src/Assets/Scripts/Main/Controller/EndfieldController.cs
@@ -24,11 +24,9 @@ namespace Main.Controller
         private bool isEnded;
 
         [Inject]
-        public EndFieldController(
-            RootInstance rootInstance
-        )
+        public EndFieldController()
         {
-            this.rootInstance = rootInstance.CheckNull();
+            this.rootInstance = RootInstance.Shared;
             endFieldImage = GameObject.Find("EndField").GetComponent<Image>();
             endFieldImage.color = new(0, 0, 0, 0);
         }
