+++ src/Assets/Scripts/Main/Controller/GameNavigation/ScreenProtocol.cs
@@ -24,11 +24,14 @@ namespace Main.Controller.GameNavigation
 
         [Inject] public ScreenController<T> Controller { get; set; }
 
-        [Inject] public RootInstance RootInstance { get; set; }
+        protected RootInstance RootInstance => RootInstance.Shared;
 
         public void OnDestroy()
         {
-            Destroy(gameObject);
+            if (gameObject != null)
+            {
+                Destroy(gameObject);   
+            }
         }
 
         public void OnCreate()
